+++ package.json
@@ -15,10 +15,6 @@
     "tsc": "tsc"
   },
   "dependencies": {
-    "eslint-config-prettier": "^6.11.0",
-    "eslint-config-wolox-react-native": "^1.0.1",
-    "eslint-plugin-import": "^2.22.0",
-    "eslint-plugin-prettier": "^3.1.4",
     "react": "16.11.0",
     "react-native": "0.62.2"
   },
@@ -38,7 +34,11 @@
     "metro-react-native-babel-preset": "^0.58.0",
     "prettier": "^2.0.5",
     "react-test-renderer": "16.11.0",
-    "typescript": "^3.8.3"
+    "typescript": "^3.8.3",
+    "eslint-config-prettier": "^6.11.0",
+    "eslint-config-wolox-react-native": "^1.0.1",
+    "eslint-plugin-import": "^2.22.0",
+    "eslint-plugin-prettier": "^3.1.4"
   },
   "jest": {
     "preset": "react-native",
